+++ mesoSPIM/config/demo_config.py
@@ -22,7 +22,7 @@ ui_options = {'dark_mode' : True, # Dark mode: Renders the UI dark if enabled
               'flip_XYZFT_button_polarity': (True, False, False, False, False), # flip the polarity of the stage buttons (X, Y, Z, F, Theta)
               'button_sleep_ms_xyzft' : (250, 0, 250, 0, 0), # step-motion buttons disabled for N ms after click. Prevents stage overshooting outside of safe limits, for slow stages.
               'window_pos': (100, 100), # position of the main window on the screen, top left corner.
-              'usb_webcam_ID': 0, # open USB web-camera (if available): 0 (first cam), 1 (second cam), ...
+              'usb_webcam_ID': 0, # open USB web-camera (if available): None,  0 (first cam), 1 (second cam), ...
                }
 
 logging_level = 'DEBUG' # 'DEBUG' for ultra-detailed, 'INFO' for general logging level