+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - during acqusition, the currently acquiring row is highlighted in **Acquisition Manager** and in **Tile Overview** windows. 
 This has to be set up in the config file with `'x_center_position'` and `'z_center_position'` parameters for stage motion.
 - Buttons for movement in horizontal plane got shorter, for more intuitive navigation.
-- current FPS shown in the progress bar
+- Current FPS shown in the progress bar.
+- Webcam window always opens at startup, empty if no camera is present in config file: `'usb_webcam_ID': 0, # open USB web-camera (if available): None,  0 (first cam), 1 (second cam), ...`
 - Tooltips were added to the navigation buttons.
 
 ### Bugfixes :bug: